Add files via upload
Please find a revised version of LeadingIndicatorTools
(1) get_and_prepare_signals has been revised to include the 0/-ve cases. I have fixed a bug in the code (at-least I thought it was) regarding the indexing, and have commented out the lines to read in data from covidcast for now. 
(2) identify_beginning_increase_points is a new function, which takes in the raw covidcast signals and then returns a list of dataframes with a column that specifies points at the beginning of a rise in cases or indicator.
(3) plot_smoothed_signals_increase_point is a new function that plots the smoothed version of the signals and highlights the points identified in (2). Some further work needs to be done on re-scaling the axes.